use multi-stage build to reduce image size
This commit changes the Dockerfile to use a multi-stage build. In the second stage, we begin with a bare debian:stretch-slim base image, and the compiled dsistudio program is copied over. Runtime dependencies are also installed. This second stage ommits unnecessary build artifacts and compile-time dependencies that are not needed at runtime. This commit reduces the overall image size (uncompressed) from 2.6 GB to 427 MB.